Version 0.3.0 (AS1 support) (#12)
* add base for as1

* change as1 login handling

* fix as1 music

* add as1 minigames

* remove RPC from development mode

* add basic items to as1

* add beta test party

* fix PSA test for as1

* add automatic account creation for as1

* add as1 commands

* add halloween 2005 rooms

* add puffle discovery party

* add christmas 2005

* add winter luau base

* add valentine's day celebration

* add base pizza parlor opening party

* add april fools party 2006

* add as1 party endings

* add as1 newspapers up to 26

* enforce proper room consistency until iceberg

* add settings menu

* add as1 catalogues

* restructure media folders

* Add "Open mods folder" button (#9)

* Go back to the game if a page can't load (#10)

* Go back to the game if a page can't load

* Change to prevent leaving localhost in the first place

* change reload setting to go to main page

* remove leftover log

* fix server error with website

* add supermanover00's winter luau town recreation

* add auto updater for default media on opening

* add auto updater for clothing media

* add setup build options

* disable the ability for medias to be changed in development

* create settings API for client and server to communicate non directly

* add year filter for timeline

* make timeline start on the year of current date

* add puffle launch and system defender

* update stadium games files

* improve visuals and presentation of some windows

* change window title

* store puffle launch data as base64 instead

* add cheat commands for puffle launch

* slightly improve commands system

* make media changes for upcoming refactor (breaks game for now)

* major refactor of the media serving system

* remove winter luau

* fix leftover winter luau music

* make mods list vertical

* change client to access the new media route

* add accurate dates for dojo courtyard swfs

* remove party music from base crumbs

* change as1 check for arbitrary 2008 date

* add mountain expedition event (credits to victando)

* add proper snow forts for mountain expedition

* remanage stage plays from around mid 2010 (underwater, squidzoid)

* add existing catalogues from mid 2010

* add news crumbs generator

* add mountain expedition news crumbs

* add july 2010 newspapers

* add proper mountain expedition stampbook

* update credits

* make puffle launch game data default null and database migration

* enable autoupdater for windows

* recode the version warning

* add linux install script

* fix linux script

* update linux script with media version

* removing client files from server

* fix server depending on client files

* update version

* add database guide

* fix database readme

* update link to installer script

---------

+++ src/client/settings.ts
@@ -1,21 +1,23 @@
 import { BrowserWindow, ipcMain, dialog } from "electron";
-import { SettingsManager } from "../server/settings";
+import electronIsDev from "electron-is-dev";
 import path from "path";
 import fs from 'fs';
-import { download } from "./download";
-import { VERSION } from "../common/version";
-import { unzip } from "./unzip";
+import { downloadMediaFolder } from "./media";
+import { GlobalSettings } from "../common/utils";
 
-export const createSettingsWindow = async (settingsManager: SettingsManager, mainWindow: BrowserWindow) => {
-  if (settingsManager.isEditting) {
+// preventing garbage collection
+let settingsWindow: BrowserWindow;
+
+export const createSettingsWindow = async (globalSettings: GlobalSettings, mainWindow: BrowserWindow) => {
+  if (globalSettings.isEditting) {
     await dialog.showMessageBox(mainWindow, {
       buttons: ['Ok'],
       title: 'Error',
       message: `You already have the settings window open.`,
     });
   } else {
-    settingsManager.isEditting = true;
-    const settingsWindow = new BrowserWindow({
+    globalSettings.isEditting = true;
+    settingsWindow = new BrowserWindow({
       width: 500,
       height: 500,
       title: "Settings",
@@ -25,43 +27,33 @@ export const createSettingsWindow = async (settingsManager: SettingsManager, mai
     });
   
     settingsWindow.loadFile(path.join(__dirname, 'views/settings.html'));
-    settingsWindow.webContents.on('did-finish-load', () => {
-      settingsWindow.webContents.send('receive-settings', settingsManager.settings);
-    });
+
+    if (electronIsDev) {
+      settingsWindow.webContents.on('did-finish-load', () => {
+        settingsWindow.webContents.openDevTools();
+      });
+    }
 
     settingsWindow.on('close', () => {
-      settingsManager.isEditting = false;
-    });
-  
-    ipcMain.on('update-settings', (e, arg) => {
-      settingsManager.updateSettings(arg);
+      globalSettings.isEditting = false;
     });
 
     ipcMain.on('download-package', (e, arg) => {
       (async () => {
-        // avoid collision (unlink only deletes after the app is closed)
-        const zipName = String(Date.now()) + '.zip'
-        const mediaDir = path.join(process.cwd(), 'media');
-        const zipDir = path.join(mediaDir, zipName)
-        const success = await download(`${arg}-${VERSION}.zip`, zipDir)
-        if (success) {
-          await unzip(zipDir, path.join(mediaDir, arg))
+        downloadMediaFolder(arg, () => {
           settingsWindow.webContents.send('finish-download', arg)
-          settingsManager.updateSettings({
-            [arg]: true
-          })
-        } else {
+        }, () => {
           settingsWindow.webContents.send('download-fail')
-        }
+        })
       })()
     })
 
     ipcMain.on('delete-package', (e, arg) => {
-      fs.rmdirSync(path.join(process.cwd(), 'media', arg), { recursive: true })
+      // must not remove packages in development, as that would greatly disturb git
+      if (!electronIsDev) {
+        fs.rmdirSync(path.join(process.cwd(), 'media', arg), { recursive: true })
+      }
       settingsWindow.webContents.send('finish-deleting', arg)
-      settingsManager.updateSettings({
-        [arg]: false
-      })
     })
 
     ipcMain.on('reload-window', () => {
